admiraldiscovery: Discover admiral Functions

Package exports a function that creates an interactive gt or reactable table to help users discover the admiral functions that are used to derive ADaM variables.

See also

Author

Maintainer: Daniel D. Sjoberg danield.sjoberg@gmail.com (ORCID)

Authors:

Amelia H. Tran tran.amelia@gene.com (ORCID)